refactor: #156 - HomeViewReactor용 구조체들 파일로 분리
Domain - HomeEntity에 생성됨
